Fix Solaris compiler warning in signal.c
git-svn-id: https://bacula.svn.sourceforge.net/svnroot/bacula/trunk@9083 91ce42f0-d328-0410-95d8-f526ca767f89

+++ bacula/src/lib/signal.c
@@ -90,7 +90,7 @@ static void dbg_print_bacula()
    char buf[512];
 
    snprintf(buf, sizeof(buf), "%s/%s.%d.bactrace", 
-            working_directory, my_name, getpid());
+            working_directory, my_name, (int)getpid());
    FILE *fp = fopen(buf, "ab") ;
    if (!fp) {
       fp = stderr;
